# stacks/python-fastapi/templates/openapi-template.yaml
openapi: 3.0.0
info:
  title: Your Project API
  description: FastAPI project with authentication and modular structure
  version: 1.0.0
  contact:
    name: API Support
    email: support@yourproject.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.yourproject.com
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Authentication
    description: User authentication and authorization
  - name: Resources
    description: Resource management
  - name: Items
    description: Item management

components:

  schemas:
    # Health Check
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: healthy
        database:
          type: boolean
          example: true
        environment:
          type: string
          example: development

    # Authentication
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - fullName
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: securepassword123
        fullName:
          type: string
          example: John Doe

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: securepassword123

    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful
        user:
          type: object
          properties:
            userId:
              type: string
              example: user-123
            email:
              type: string
              example: user@example.com
            fullName:
              type: string
              example: John Doe

    UserResponse:
      type: object
      properties:
        userId:
          type: string
          example: user-123
        email:
          type: string
          example: user@example.com
        fullName:
          type: string
          example: John Doe
        permissions:
          type: array
          items:
            type: string
          example: ["read", "write"]

    # Resources
    CreateResourceRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: My First Resource
        description:
          type: string
          minLength: 1
          maxLength: 10000
          example: This is the description of my first resource.

    ResourceResponse:
      type: object
      properties:
        resourceId:
          type: string
          example: resource-123
        name:
          type: string
          example: My First Resource
        description:
          type: string
          example: This is the description of my first resource.
        ownerId:
          type: string
          example: user-123
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z

    ResourceListResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ResourceResponse'
        totalCount:
          type: integer
          example: 100
        hasMore:
          type: boolean
          example: true

    UpdateResourceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Updated Resource Name
        description:
          type: string
          minLength: 1
          maxLength: 10000
          example: Updated description

    # Comments
    CreateCommentRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
          example: Great post!

    CommentResponse:
      type: object
      properties:
        commentId:
          type: string
          example: comment-456
        postId:
          type: string
          example: post-123
        content:
          type: string
          example: Great post!
        authorId:
          type: string
          example: user-123
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z

    CommentListResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
        totalCount:
          type: integer
          example: 50

    # Error Responses
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Resource not found

paths:
  /health:
    get:
      summary: Health check
      description: Check API health and database connectivity
      tags:
        - Health
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/auth/register:
    post:
      summary: Register new user
      description: Create a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /api/auth/me:
    get:
      summary: Get current user
      description: Get authenticated user information
      tags:
        - Authentication
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /api/posts:
    post:
      summary: Create post
      description: Create a new blog post
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '200':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List posts
      description: Get list of blog posts with pagination
      tags:
        - Posts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'

  /api/posts/{postId}:
    get:
      summary: Get post
      description: Get a single blog post by ID
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update post
      description: Update an existing blog post
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete post
      description: Delete a blog post
      tags:
        - Posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Post deleted successfully
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/posts/{postId}/comments:
    post:
      summary: Create comment
      description: Add a comment to a blog post
      tags:
        - Comments
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '200':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List comments
      description: Get all comments for a blog post
      tags:
        - Comments
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'